Text.Html.Entity

Entity lookup

entityCodePoints :: Text -> Maybe [Int]

Given an entity looks up unicode code points that represent it.

>>> entityCodePoints "&mu;"
Just [966]

>>> entityCodePoints "&nano;"
Nothing

entityChars :: Text -> Maybe Text

Given an entity returns it's textual representation (if any).

>>> entityChars "&mu;"
Just "μ"

>>> entityChars "&nano;"
Nothing

Entity validation

isValidEntity :: Text -> Bool

Check if entity is valid (if such entity exists).

>>> isValidEntity "&mu;"
True

>>> isValidEntity "&nano;"
False

Entity lookup by name

entityNameCodePoints :: Text -> Maybe [Int]

Given a name of entity looks up code points that represent it.

>>> entityNameCodePoints "mu"
Just [966]

>>> entityNameCodePoints "nano"
Nothing

entityNameChars :: Text -> Maybe Text

Given a name of entity returns it's textual representation (if any).

>>> entityNameChars "mu"
Just "μ"

>>> entityNameChars "nano"
Nothing

Entity name validation

isValidEntityName :: Text -> Bool

Check if entity name is valid (if corresponding entity exists).

>>> isValidEntityName "mu"
True

>>> isValidEntityName "nano"
False